% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{add_cluster_assignment}
\alias{add_cluster_assignment}
\title{Add cluster with iterative variable selection}
\usage{
add_cluster_assignment(
  df,
  v_cluster = NULL,
  k = NULL,
  maxit = 10,
  return_df_cluster_instead = FALSE,
  new_var_name = "Cluster"
)
}
\arguments{
\item{df}{data.frame}

\item{v_cluster}{(character) vector of variable names to use in clustering (if \code{NULL}, use all)}

\item{k}{(numeric) number of clusters}

\item{maxit}{(numeric) maximum number of iterations to select significant outcomes of clusters}

\item{return_df_cluster_instead}{(logical) whether to return \code{df} with only final clustering
variables and cluster assignment (default \code{FALSE})}

\item{new_var_name}{(character) new variable name}
}
\value{
\code{df} with new cluster variable added
}
\description{
This function adds cluster assignment to each row of a data.frame with
iterative variable selection during clustering.
}
\details{
An initial set of clusters is determined using \code{\link[=get_cluster]{get_cluster()}} (Gower distance matrix with
clustering around medoids). Then, significant differences between clusters are determined using
\code{\link[=get_sig_differences_between_groups]{get_sig_differences_between_groups()}} (ANOVA for numeric variables, Chi-square for categorical
variables), and clustering is performed again using the reduced set of significant variables.
This variable-reduction method is repeated until all remaining variables are significant,
or until \code{maxit} is reached. If \code{maxit} is reached before convergence, a warning is
thrown.

When \code{maxit} is set to 0, a warning is thrown indicating that no variable selection
is performed, and this is equivalent to using \code{\link[=get_cluster]{get_cluster()}} directly.
}
\examples{
\dontrun{
df <- df |> add_cluster_assignment()

df |>
   add_cluster_assignment(return_df_cluster_instead=TRUE) |>
   plot_density_by_groups(group="Cluster")
}
}
