% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{add_group_key_to_desc_tbl}
\alias{add_group_key_to_desc_tbl}
\title{Adds group key to table generated by some tbl_fn()}
\usage{
add_group_key_to_desc_tbl(tbl, group_key)
}
\arguments{
\item{tbl}{Table generated by some tbl_fn()}

\item{group_key}{From \code{dplyr::group_map()}}
}
\value{
Table with group key added
}
\description{
This function adds the group key to a table generated by some \code{tbl_fn()}. This is
used by \code{get_desc_table()} when a grouping variable is specified.
}
