% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{add_predictions_from_automl}
\alias{add_predictions_from_automl}
\title{Add predictions from an AutoML object to a test set}
\usage{
add_predictions_from_automl(automl, tt, tt_h2o, add_confidence_level = TRUE)
}
\arguments{
\item{automl}{H2O AutoML object (from \code{\link[h2o:h2o.automl]{h2o::h2o.automl()}} or \code{\link[=get_automl_model]{get_automl_model()}})}

\item{tt}{train-test list (see \code{\link[=ttsplit]{ttsplit()}})}

\item{tt_h2o}{H2O pointers to train and test frames (see \code{\link[=ship_train_and_test_to_h2o]{ship_train_and_test_to_h2o()}})}

\item{add_confidence_level}{(logical) whether to add confidence levels
alongside predictions in test set (default \code{TRUE})}
}
\value{
train-test list with predictions (and optionally confidence levels) added to \code{$test}
}
\description{
This function generates predictions from an AutoML object for the test
set in a train-test split, and adds the predictions to the test set.
Confidence levels can also optionally be added alongside predictions.
}
\examples{
\dontrun{
init_h2o()
tt <- df |> ttsplit()
tt_h2o <- tt |> ship_train_and_test_to_h2o()
automl <- tt_h2o |> get_automl_model(target="y1")
tt <- add_predictions_from_automl(automl)
}

}
\seealso{
\code{\link[h2o:h2o.predict]{h2o::h2o.predict()}}, \code{\link[h2o:h2o.automl]{h2o::h2o.automl()}}, \code{\link[=get_automl_model]{get_automl_model()}}
}
\concept{machine_learning}
