% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{add_target_back_to_test_set_from_ref_table}
\alias{add_target_back_to_test_set_from_ref_table}
\title{Add target values back in the test set}
\usage{
add_target_back_to_test_set_from_ref_table(tt, ref_table)
}
\arguments{
\item{tt}{train-test list (see \code{\link[=ttsplit]{ttsplit()}})}

\item{ref_table}{reference table (see Details)}
}
\value{
\code{tt} with target replaced with original values in \code{$test}
}
\description{
This function adds back the target values in the test set using
a reference table.
}
\details{
The reference table is a data.frame initially created by
\code{\link[=remove_target_from_test_and_add_ref_to_env]{remove_target_from_test_and_add_ref_to_env()}}.
}
\note{
This function implements the method described in
\href{https://en.d22consulting.com/quantcafe/preventing-target-leakage}{Preventing Target Leakage} (D22 QuantCafé, 2021).
}
\examples{
\dontrun{
tt <- tt |> add_target_back_to_test_set_from_ref_table(target_values)
}

}
\seealso{
\code{\link[=remove_target_from_test_and_add_ref_to_env]{remove_target_from_test_and_add_ref_to_env()}}
}
\concept{machine_learning}
