% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{add_vars_from_one_df_to_another}
\alias{add_vars_from_one_df_to_another}
\title{Add variables from one data.frame to another}
\usage{
add_vars_from_one_df_to_another(df_to, df_from, by_to, by_from = NULL, vars)
}
\arguments{
\item{df_to}{data.frame in which to add the wanted variables}

\item{df_from}{data.frame that contains the wanted variables}

\item{by_to}{(character) matching variable in \code{df_to}}

\item{by_from}{(character, optional) matching variable in \code{df_from}
(if different from \code{by_to})}

\item{vars}{(character) vector of variable names to add to \code{df_to}}
}
\value{
\code{df_to} with columns added (if any row match by \code{by})
}
\description{
This function adds some variables from one data.frame to another,
by matching rows by a given variable.
}
\details{
This is a convenience function using \code{merge()}, with only selected variables merged
into the first data.frame, and with the order of columns preserved. The matching variable
may or may not be named the same in both data.frames (as controlled by \code{by_to} and
\code{by_from}).
}
\examples{
\dontrun{
df <- df |> add_vars_from_one_df_to_another(ref_table, by_to="id", vars=c("Gender", "Age"))
}

}
\seealso{
\code{\link[base:merge]{base::merge()}}
}
