% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{ana_fn_aov}
\alias{ana_fn_aov}
\title{Returns p-values from ANOVA for specified vars and grouping variable}
\usage{
ana_fn_aov(
  df,
  vars,
  group,
  add_statistic = FALSE,
  add_posthoc = FALSE,
  posthoc_adjust = "none"
)
}
\arguments{
\item{df}{data.frame}

\item{vars}{Vector of variable names}

\item{group}{Name of grouping variable}

\item{add_statistic}{(logical) Whether to add F-test to table}

\item{add_posthoc}{(logical) Whether to add post-hoc pairwise comparisons to table}

\item{posthoc_adjust}{Method for adjusting p-values (see \code{\link[stats:p.adjust]{stats::p.adjust()}}; default 'none')}
}
\value{
\code{data.frame} with columns \code{Var}, \code{p} (formatted p-values), \code{F-test} (if requested),
and (possibly adjusted) p-values from pairwise comparisons (if requested)
}
\description{
This function computes p-values using ANOVA for each variable in \code{vars} and grouping variable
specified in \code{group}
}
\details{
This should be passed as \code{ana_fn} when calling \code{\link[=get_desc_table]{get_desc_table()}} with a \code{group} argument
to obtain p-values. If you want to create a custom \code{ana_fn}, you should model it after this one.
}
\examples{
\dontrun{
df |> ana_fn_aov("age", "TreatmentGroup")
}

}
\seealso{
\code{\link[stats:aov]{stats::aov()}}, \code{\link[=get_desc_table]{get_desc_table()}}
}
\concept{descriptives}
