% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{ana_fn_rm_aov}
\alias{ana_fn_rm_aov}
\title{Returns p-values from repeated-measures (RM) ANOVA for specified vars and grouping (time) variable}
\usage{
ana_fn_rm_aov(df, vars, group, id, add_cohen = FALSE)
}
\arguments{
\item{df}{data.frame (in long format)}

\item{vars}{Vector of variable names}

\item{group}{Name of grouping (time) variable}

\item{id}{ID var (unique identifier)}

\item{add_cohen}{(logical) Whether to add Cohen's d in addition to p-values}
}
\value{
\code{data.frame} with \code{Var} and \code{p} (formatted p-values) columns
}
\description{
This function computes p-values using RM-ANOVA for each variable in \code{vars} and grouping (time) variable
specified in \code{group}
}
\details{
This should be passed as \code{ana_fn} when calling \code{get_desc} with \code{group} and \code{id}
arguments to obtain p-values. Alternatively, this is the \code{ana_fn} used in wrapper
\code{get_desc_time} when data.frame is not yet in long format.
}
\examples{
df |>
   to_long(list(score=c("y1", "y2"))) |>
   ana_fn_rm_aov(vars="score", group="Time", id="id", add_cohen=TRUE)

}
\seealso{
\code{\link[=to_long]{to_long()}}, \code{\link[=get_desc_time]{get_desc_time()}}, \code{\link[stats:aov]{stats::aov()}}
}
