% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{calc_cluster_importances}
\alias{calc_cluster_importances}
\title{Calculate variable importances in clustering}
\usage{
calc_cluster_importances(df, nrep = 100, ...)
}
\arguments{
\item{df}{data.frame}

\item{nrep}{(positive int) number of repetitions in the simulation (default \code{100})}

\item{...}{arguments passed to \code{add_cluster_assignment()}}
}
\value{
(table) percentage of times each variable was used in the final clustering step
}
\description{
This function calculates variable importances when clustering with variable selection.
}
\details{
Variable importances are estimated through simulation with \code{nrep} calls to \code{add_cluster_assignment()}
with the specified configuration via \code{...}. The function retrieves the number of times each
variable was kept at the final step of clustering.
}
\examples{
\dontrun{
df |>
  calc_cluster_importances(k=2,
                           maxit=100,
                           elimination="bidirectional",
                           max_vars_rm_or_add_each_it=1)
}
}
\seealso{
\code{\link[=add_cluster_assignment]{add_cluster_assignment()}}
}
\concept{cluster}
