% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{compare_pairs_of_vars}
\alias{compare_pairs_of_vars}
\title{Compares pairs of variables using matched-pairs t-tests}
\usage{
compare_pairs_of_vars(df, vars, order_output = TRUE)
}
\arguments{
\item{df}{data.frame}

\item{vars}{Vector of variable names}

\item{order_output}{(logical) Whether to rank-order row and columns of subsets matrix by mean size (recommended)}
}
\value{
List with two matrices:
\describe{
\item{p_uncorrected}{Matrix of (unadjusted) p-values}
\item{subsets}{Matrix indicating whether each pair of variables are
part of the same subset (indicated by "SAMESET") or are instead significantly
different (blank entry) following a Bonferroni correction}
}
}
\description{
This function compares pairs of variables using matched-pairs t-tests,
and also provides "homogeneous subsets" (similar to SPSS) following
a Bonferroni correction.
}
\details{
p-values come from matched-pairs t-tests assuming equal variances
and are uncorrected for the number of tests. Subsets are generated following
a Bonferroni correction, and the rows and columns of the resulting matrix can
be ordered by variable mean: This is recommended to simplify examination
of the homogeneous subsets.
}
\examples{
df |> compare_pairs_of_vars(c("x1", "x2", "x3"))
}
