% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{compare_sig_effects_in_two_pub_tables}
\alias{compare_sig_effects_in_two_pub_tables}
\title{Compare significant effects from two publication tables}
\usage{
compare_sig_effects_in_two_pub_tables(table1, table2, pattern = "*", fixed = T)
}
\arguments{
\item{table1}{A publication table}

\item{table2}{Another publication table}

\item{pattern}{Pattern used to determine significant effects (default is *)}

\item{fixed}{(logical) Whether pattern is to be searched as is (TRUE)
or as a regular expression (FALSE)}
}
\description{
This function prints to console which effects are significant in table1 but not in table2,
and which are significant in table2 but not in table1. The two tables are
publication tables generated by any of the \code{make_pub_table_from_...()} functions
}
\examples{
library(lavaan)
ana <- list()
ana$y1 <- sem("y1 ~ x1+x2", df, missing="ml.x")
ana$y2 <- sem("y2 ~ x1+x2", df, missing="ml.x")
pub_sem <- ana |> make_pub_table_from_lavaan_models()

library(broom)
ana <- list()
ana$y1 <- lm(y1 ~ x1+x2, df) |> tidy()
ana$y2 <- lm(y2 ~ x1+x2, df) |> tidy()
pub_lm <- ana |> make_pub_table_from_broom_tidy()

# sensitivity analysis: effect of missing data
pub_lm |> compare_sig_effects_in_two_pub_tables(pub_sem)

}
\concept{models}
