% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{contr.dummy_or_effect}
\alias{contr.dummy_or_effect}
\title{Contrasts for factors using effect coding (when k > 2) or dummy coding (when k = 2)}
\usage{
contr.dummy_or_effect(...)
}
\arguments{
\item{...}{Passed to \code{contr.treatment()}}
}
\value{
Matrix with contrasts
}
\description{
Contrasts for factors using effect coding (when k > 2) or dummy coding (when k = 2)
}
\examples{
contr.dummy_or_effect(3)
options(contrasts=c(unordered="contr.dummy_or_effect", ordered="contr.poly"))

}
\concept{data_prep}
