% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{copy2}
\alias{copy2}
\title{Writes table to text file in current directory}
\usage{
copy2(tbl)
}
\arguments{
\item{tbl}{Object to be written}
}
\description{
This is a wrapper for \code{write.table()} with \code{file="gentleman_out.txt"}.
}
\examples{
\dontrun{
df |> copy2()
}

}
\concept{helpers}
