% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{copy3}
\alias{copy3}
\title{Wrapper for writing table to clipboard}
\usage{
copy3(tbl)
}
\arguments{
\item{tbl}{Object to be written}
}
\description{
This is a wrapper for \code{write.table()} with \code{file="clipboard-9999"}
and \code{quote=FALSE}. Ideal for copy-pasting a data.frame into a text file.
}
\examples{
df |> copy3()

}
\concept{helpers}
