% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{decompose_interaction}
\alias{decompose_interaction}
\title{Decompose 2-way (x1*x2) interaction}
\usage{
decompose_interaction(
  model,
  df = NULL,
  x1,
  x2,
  at_x1 = NULL,
  at_x2 = NULL,
  ci = FALSE
)
}
\arguments{
\item{model}{A model object compatible with \code{emmeans} (e.g. \code{lm})}

\item{df}{data.frame or NULL if \code{model} has \code{model$model}}

\item{x1, x2}{(character) variable name}

\item{at_x1, at_x2}{(numeric or character vector) \code{x1} and \code{x2} values at which to calculate
and compare means (if \code{NULL}, all unique values)}

\item{ci}{(logical) Whether to plot confidence intervals}
}
\value{
List with elements:
\describe{
\item{slopes}{Simple slopes and pairwise comparisons (from calling \code{emmeans::emtrends()})}
\item{means}{(adjusted) means for every combination of \code{x1*x2}
(may need to specify \code{at_x1} and/or \code{at_x2} values)}
\item{contrasts}{Comparisons of (adjusted) \code{x1} means at each value of \code{x2}
(may need to specify \code{at_x1} and/or \code{at_x2} values)}
\item{plot}{Plot (\code{ggplot2} line graph using data from \code{emmeans::emmip()})}
}
}
\description{
This function decomposes a 2-way x1*x2 interaction into simple effects,
and provides means, pairwise comparisons, and a plot.
}
\details{
This function uses the facilities provided by the library \pkg{emmeans}, and
follows the procedure outlined by
\href{https://stats.oarc.ucla.edu/r/seminars/interactions-r/}{UCLA OARC Statistical and Data Analytics}.
It might be helpful to execute the function with x1 and x2 inverted, and see which
set of results and plot make most sense for your scenario.

Note that if \code{x1} or \code{x2} has many different distinct values and \code{at_x1}
or \code{at_x2} is not specified, then not all elements of the returned list will be useful
(\code{$means} and \code{$contrasts} in particular may be difficult to use).
}
\examples{
\dontrun{
fit <- lm(SelfEsteem ~ Gender*AttitudeTowardsSchool, df)
fit |> decompose_interaction(x1="Gender", x2="AttitudeTowardsSchool")
}

}
\seealso{
\code{\link[emmeans:emmeans]{emmeans::emmeans()}}, \code{\link[emmeans:emtrends]{emmeans::emtrends()}}, \code{\link[emmeans:emmip]{emmeans::emmip()}}
}
\concept{models}
