% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{dichotomize}
\alias{dichotomize}
\title{Dichotomize by picking value returned by a function}
\usage{
dichotomize(df, vars, value_fn = max, na.rm = TRUE, ...)
}
\arguments{
\item{df}{data.frame}

\item{vars}{(character) variable names to dichotomize}

\item{value_fn}{function that returns desired value for \code{1} (default \code{max()})}

\item{na.rm}{(logical) whether to remove NAs before applying \code{value_fn} (default \code{TRUE})}

\item{...}{(optional) additional arguments passed to \code{value_fn}}
}
\value{
\code{df} with \code{vars} dichotomized
}
\description{
This function dichotomizes (0/1) a variable by applying a function to it
and checking whether each element equals the function's return value.
}
\examples{
df <- df |> dichotomize(vars="x1")

}
\concept{data_prep}
