% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{empty_df}
\alias{empty_df}
\title{Create empty data.frame with column names}
\usage{
empty_df(colnames)
}
\arguments{
\item{colnames}{(character) vector of column names}
}
\value{
Empty \code{data.frame} with column names \code{colnames}
}
\description{
This function creates an empty \code{data.frame} with pre-defined column names
}
\examples{
df <- empty_df("x"\%p\%1:10)

}
\concept{helpers}
