% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{format_p}
\alias{format_p}
\title{Format p-values with 3 decimals, no leading 0s, and < .001}
\usage{
format_p(p)
}
\arguments{
\item{p}{A (named) numeric vector of p-values}
}
\value{
A (named) character vector of formatted p-values
}
\description{
This function formats a vector of p-values to match the common requirement of 3 decimals with
no leading zeros and indicated as "< .001" rather than .000.
}
\examples{
format_p(c(0.000007, 0.15))
}
