% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_all_pairs}
\alias{get_all_pairs}
\title{Get all possible pairs from vector}
\usage{
get_all_pairs(vec, direction = c("direct", "double", "reverse"))
}
\arguments{
\item{vec}{Vector}

\item{direction}{\arguments{
\item{\code{direct} (default)}
{Like permutation: (a,b) and (b,a) are treated as the same pair and only one is returned}
\item{\code{double}}
{Like combination: (a,b) and (b,a) are treated as distinct pairs}
\item{\code{reverse}}
{Like direct, but the two columns are reversed}

}}
}
\value{
Two-column matrix of pairs
}
\description{
This function returns a matrix of all pairs of values in a vector.
}
\examples{
get_all_pairs(c("x1", "x2", "y1", "y2"))
}
