% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_all_pairs_with_op}
\alias{get_all_pairs_with_op}
\title{Helper function to generate pairs with an operator separator}
\usage{
get_all_pairs_with_op(vars, op, ...)
}
\arguments{
\item{vars}{List of variables}

\item{op}{Operator (typically a lavaan operator like \code{~})}

\item{...}{Named arguments passed to \code{get_all_pairs()}}
}
\value{
Character vector with all pairs with operator separator
}
\description{
Used for generating formulas in lavaan models.
}
\concept{helpers}
