% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{get_automl_model}
\alias{get_automl_model}
\title{Run an H2O AutoML model}
\usage{
get_automl_model(config, tt_h2o, ...)
}
\arguments{
\item{config}{configuration object (see \code{\link[=get_automl_config]{get_automl_config()}})}

\item{tt_h2o}{list of H2O pointers to H2O frames (see \code{\link[=ship_train_and_test_to_h2o]{ship_train_and_test_to_h2o()}})}

\item{...}{additional arguments passed to \code{\link[h2o:h2o.automl]{h2o::h2o.automl()}} (see Details)}
}
\value{
AutoML object from H2O
}
\description{
This function is a wrapper for \code{\link[h2o:h2o.automl]{h2o::h2o.automl()}},
with relevant options specified through the configuration
object generated by \code{\link[=get_automl_config]{get_automl_config()}}.
}
\details{
Only \code{$train} is currently used in \code{tt_h2o}.

Additional arguments passed to H2O via \code{...} may not include the following,
which have fixed values here:
\itemize{
\item \code{stopping_tolerance}: .01
\item \code{keep_cross_validation_predictions}: \code{FALSE}
\item \code{keep_cross_validation_models}: \code{FALSE}
}
}
\examples{
\dontrun{
init_h2o()
tt <- df |> ttsplit()
tt_h2o <- tt |> ship_train_and_test_to_h2o()
config <- get_automl_config(target="y1")
automl <- config |> get_automl_model(tt_h2o)
}

}
\seealso{
\code{\link[=get_automl_config]{get_automl_config()}}, \code{\link[h2o:h2o.automl]{h2o::h2o.automl()}}
}
