% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{get_cluster}
\alias{get_cluster}
\title{Get cluster assignment using Gower distance matrix}
\usage{
get_cluster(df, v_cluster = NULL, k = NULL, weights = NULL)
}
\arguments{
\item{df}{data.frame}

\item{v_cluster}{variables used to compute Gower distances between rows (if \code{NULL}, use all)}

\item{k}{number of clusters (if \code{NULL}, determined optimally; see Details)}

\item{weights}{(numeric vector) variable weights for calculating Gower distances (default all 1)}
}
\value{
(factor) vector of cluster assignments (0 to k-1)
}
\description{
This function assigns each row of a data.frame to a cluster based on the Gower distance matrix,
and either a pre-specified or an optimal number of clusters.
}
\details{
First, a distance matrix is computed using \code{\link[cluster:daisy]{cluster::daisy()}} with \code{metric="gower"}
and \code{stand=TRUE}. Next, clustering is performed around medoids (a more robust version of k-means
clustering) as implemented in \code{\link[cluster:pam]{cluster::pam()}}.

If no number of clusters \code{k} was specified, then the optimal
number of clusters is determined for the current distance matrix using \code{\link[NbClust:NbClust]{NbClust::NbClust()}} with the
\code{method="median"} and \code{index="silhouette"}.
}
\examples{
df |> get_cluster() |> table()

}
\concept{cluster}
