% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_crosslagged_model}
\alias{get_crosslagged_model}
\title{Generate lavaan syntax for cross-lagged model}
\usage{
get_crosslagged_model(vars_list, random_intercepts = FALSE)
}
\arguments{
\item{vars_list}{List of vectors of variable names of length t (number of repeated assessments)}

\item{random_intercepts}{(logical) Whether to add random-intercepts (RI-CLPM model)}
}
\value{
Character value to be used with lavaan as model syntax
}
\description{
This function generates the lavaan syntax for a cross-lagged panel model (CLPM).
With random-intercepts, the syntax yields the random-intercepts cross-lagged panel
model (RI-CLPM). The function takes an arbitrary number of different variables,
and an arbitrary number of repeated assessments.
}
\examples{
library(lavaan)
vars <- list(
   c("x1", "x2", "x3"),
   c("y1", "y2", "y3")
)
get_crosslagged_model(vars) |>
   sem(df) |>
   summary()
}
