% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{get_desc_table}
\alias{get_desc_table}
\title{Get a table of descriptive statistics for numeric or factor variables}
\usage{
get_desc_table(df, vars, tbl_fn, group = NULL, ana_fn = NULL, ...)
}
\arguments{
\item{df}{data.frame}

\item{vars}{Vector of variable names}

\item{tbl_fn}{Function that generates table}

\item{group}{Name of grouping variable}

\item{ana_fn}{Function that computes p-values (or other statistics to compare groups)}

\item{...}{Additional named arguments passed to \code{ana_fn}}
}
\value{
Table of descriptive statistics
}
\description{
This function returns a table of descriptive statistics generated by some \code{tbl_fn} function.
Additionally, statistics can be computed separately by groups. Groups can also be compared using
some \code{ana_fn} function.
}
\details{
For numeric variables, a natural choice is to use \code{tbl_fn=tbl_fn_num}, while for
categorical variables, use \code{tbl_fn=tbl_fn_fac}. Alternatively, custom functions can be specified.

If \code{group} is specified, you have the option of requesting p-values (or some other statistic)
comparing groups. For numeric variables, a natural choice is \code{ana_fn=ana_fn_aov}, which
computes p-values using analysis of variance. For factor variables, p-values can be obtained
with Chi-square tests using \code{ana_fn=ana_fn_chisq}.
}
\examples{
df |> get_desc_table(
    vars=c("age", "score"),
    tbl_fn=tbl_fn_num,
    group="TreatmentGroup",
    ana_fn=ana_fn_aov
)

df |> get_desc_table(
    vars=c("Nationality", "Diagnostic"),
    tbl_fn=tbl_fn_fac,
    group="TreatmentGroup",
    ana_fn=ana_fn_chisq
)

}
\seealso{
\code{\link[=get_desc_time]{get_desc_time()}}, \code{\link[=tbl_fn_num]{tbl_fn_num()}}, \code{\link[=tbl_fn_fac]{tbl_fn_fac()}}
\code{\link[=ana_fn_aov]{ana_fn_aov()}}, \code{\link[=ana_fn_chisq]{ana_fn_chisq()}}
}
