% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{get_desc_time}
\alias{get_desc_time}
\title{Get a table of descriptive statistics using RM-ANOVA for numeric variables}
\usage{
get_desc_time(df, repeated_vars, ...)
}
\arguments{
\item{df}{data.frame}

\item{repeated_vars}{Named list of variables to be pivoted to long format}

\item{...}{Additional named arguments passed to \code{ana_fn_rm_aov()}}
}
\value{
Table of descriptive statistics (output from \code{get_desc_table()})
}
\description{
This function returns a table of descriptive statistics for numeric variables
repeated over time (or otherwise repeated within subjects), with p-values calculated
through repeated-measures ANOVA. The data.frame is provided in wide format.
}
\details{
This function is a wrapper, pivoting a data.frame to long format with \code{to_long()}
before calling \code{get_desc_table()} with \code{tbl_fn=tbl_fn_num}, \code{group="Time"},
\code{ana_fn=ana_fn_rm_aov}, and \code{id="id"}. Currently there is no way to change
the analysis function.
}
\examples{
v_repeated <- list(
   x=c("x1","x2","x3"),
   y=c("y1", "y2", "y3")
   )
df |> get_desc_time(v_repeated)

}
\seealso{
\code{\link[=get_desc_table]{get_desc_table()}}, \code{\link[=to_long]{to_long()}}, \code{\link[=ana_fn_rm_aov]{ana_fn_rm_aov()}}
}
