% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_fmodel}
\alias{get_fmodel}
\title{Get model object(s) using formula interface}
\usage{
get_fmodel(df, dv, preds, preds_link = "+", ana_fn, ...)
}
\arguments{
\item{df}{data.frame}

\item{dv}{(character) vector of outcome variable names}

\item{preds}{(character) vector of predictor variable names}

\item{preds_link}{link between predictors in model formula (default \code{+})}

\item{ana_fn}{analysis function (for example: lm, lavaan::sem)}

\item{...}{additional arguments passed to \code{ana_fn}}
}
\value{
Named list of objects returned by \code{ana_fn} with names \code{dv}
}
\description{
This function generates a model formula using a vector of predictors
and a predictor link, and returns a list of models obtained from
calling a specified analysis function with the generated formula.
Additional arguments can also be passed to the analysis function.
}
\details{
The formula is constructed by surrounding each variable with backticks
to allow for special characters in variable names. The analysis function receives
the formula and the data.frame as first and second (positional) arguments,
respectively.
}
\examples{
df |>
   get_fmodel(dv="y3", preds=c("x1","x2","x3"), ana_fn=lm) |>
   lapply(summary)

}
\concept{models}
