% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{get_git_commit}
\alias{get_git_commit}
\title{Retrieve current git commit information}
\usage{
get_git_commit(dir = getwd())
}
\arguments{
\item{dir}{path to repository directory (default current directory)}
}
\value{
list with elements:
\describe{
\item{HEAD}{Contents of HEAD}
\item{SHA}{Unique commit identifier}
\item{COMMIT_MSG}{Commit message}
}
}
\description{
This function retrieves the information about the current (latest)
git commit information in a specified repository directory.
}
\examples{
\dontrun{
get_git_commit()
}

}
\seealso{
\code{\link[=get_logger]{get_logger()}}, \code{\link[=logtext]{logtext()}}
}
