% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_lavaan_interaction_plot}
\alias{get_lavaan_interaction_plot}
\title{Get interaction plot for lavaan model}
\usage{
get_lavaan_interaction_plot(model, x, w, y, values_at = NULL)
}
\arguments{
\item{model}{\pkg{lavaan} object}

\item{x}{(character) variable name to plot on the \emph{x} axis}

\item{w}{(character) variable name to plot on separate lines (moderator)}

\item{y}{(character) variable name indicating the outcome}

\item{values_at}{(named list) values at which to plot \code{x} and \code{w}
(default -1, 0, and 1 for both)}
}
\value{
\code{ggplot2} plot
}
\description{
This function generates an interaction plot from a \pkg{lavaan} model
with a 2-way (x*w) interaction.
}
\details{
The predicted values are calculated using \code{\link[lavaan:lavPredictY]{lavaan::lavPredictY()}} for the
variable given in \code{y}, and observed variables are extracted from the
\code{model} object using \code{\link[lavaan:lavNames]{lavaan::lavNames()}}.
}
\examples{
library(lavaan)
"y2 ~ x1 + x2 + x1:x2" |>
   sem(df) |>
   get_lavaan_interaction_plot(x="x1", w="x2", y="y2")

}
\concept{models}
