% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_mediation_model_2level}
\alias{get_mediation_model_2level}
\title{Generate lavaan syntax for 2-level mediation model}
\usage{
get_mediation_model_2level(x, m, y)
}
\arguments{
\item{x, m, y}{(character) variable names}
}
\value{
Character value to be used with \pkg{lavaan} as model syntax
}
\description{
This function generates the \pkg{lavaan} syntax for a two-level mediation model.
The model takes one predictor (x), one mediator (m), and one outcome (y).
The mediation model is reproduced at the within level (level 1) and the
between level (level 2). A test of the difference between the level-1 and level-2
indirect effects is also provided.
}
\details{
This model is for fully-nested designs and is described in Lachowicz, Sterba, & Preacher (2015).
}
\examples{
\dontrun{
library(lavaan)
get_mediation_model_2level("x1", "x2", "x3") |>
   sem(df, group="id") |>
   summary()
}

}
\references{
Lachowicz, M. J., Sterba, S. K., & Preacher, K. J. (2015).
Investigating multilevel mediation with fully or partially nested data.
\emph{Group Processes & Intergroup Relations}, \emph{18}, 274–289.
\url{https://doi.org/10.1177/1368430214550343}.
}
\concept{models}
