% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_mediation_model_3wave}
\alias{get_mediation_model_3wave}
\title{Generate lavaan syntax for 3-wave longitudinal mediation model}
\usage{
get_mediation_model_3wave(
  x1,
  m1,
  m2,
  y1,
  y2,
  y3,
  add_resid_correl = FALSE,
  covariates = NULL
)
}
\arguments{
\item{x1}{(character) variable name for predictor at time 1}

\item{m1, m2}{(character) variable names for mediator at times 1-2}

\item{y1, y2, y3}{(character) variable names for outcome at times 1-3}

\item{add_resid_correl}{(logical) whether to add residual correlations between variables
assessed at the same time (default \code{FALSE})}

\item{covariates}{(character) covariate (control) variable names (default \code{NULL})}
}
\value{
Character value to be used with \pkg{lavaan} as model syntax
}
\description{
This function generates the \pkg{lavaan} syntax for a longitudinal mediation model
with three waves. The model uses a predictor x at time 1 (\code{x1}), one mediator at
times 2 and 3 (\code{m2}, \code{m3}), and one outcome at times 1, 2, and 3 (\code{y1}, \code{y2}, \code{y3}).
}
\details{
The indirect effect is computed as the effect from \code{x1} to \code{m2} to \code{y3}. The effect
of \code{x1} on \code{m2} is estimated after removing the effects of \code{m1} and \code{y1} on \code{m2}, and
the effect of \code{m2} on \code{y3} is estimated after removing the effects of \code{y1}, \code{y2},
\code{x1}, and \code{m1} on \code{y3}. This corresponds to the model described in Hayes (2022),
Figure 14.4(B), p. 570.

Optionally, residual correlations can be estimated between variables assessed
at the same time (as is conventional in cross-lagged models). It is also
possible to adjust for covariates, which is achieved by partialling
out the effect of the control variables given in \code{covariates} from \code{x1},
\code{m2}, and \code{y3} (the variables involved in the indirect effect). This is
the method used by JASP v0.16.4 to adjust for covariates in mediation models.
}
\examples{
\dontrun{
library(lavaan)
get_mediation_model_3wave("x1", "m1", "m2", "y1", "y2", "y3") |>
   sem(df) |>
   summary()
}

}
\references{
Hayes, A. F. (2022).
\emph{Introduction to mediation, moderation, and conditional process analysis} (3rd ed.).
New York: Guilford.

JASP Team (2022). \emph{JASP} (Version 0.16.4). Computer software:
\url{https://jasp-stats.org/}
}
\concept{models}
