% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_moderated_mediation_model}
\alias{get_moderated_mediation_model}
\title{Generate lavaan syntax for moderated mediation}
\usage{
get_moderated_mediation_model(
  x,
  m,
  y,
  mod_a = NULL,
  mod_b = NULL,
  mod_c = NULL,
  values_at = NULL,
  cov = NULL,
  adjust_exogenous = TRUE
)
}
\arguments{
\item{x, m, y}{(character) variable name(s)}

\item{mod_a, mod_b, mod_c}{(character) variable name(s) for moderators of
\emph{a} (\code{x -> m}), \emph{b} (\code{m -> y}), and \emph{c} (\code{x -> y}) paths (default \code{NULL})}

\item{values_at}{list of numeric vectors with values at which to evaluate
the conditional effects; names correspond to \code{mod_a}, \code{mod_b}, and \code{mod_c}
(default \code{NULL})}

\item{cov}{(character) variable name(s) of control variables (default \code{NULL})}

\item{adjust_exogenous}{(logical) adjust exogenous variables (\code{x}, mods)
for the effect of covariates in \code{cov}? (default \code{TRUE})}
}
\value{
Character value to be used with \pkg{lavaan} as model syntax
}
\description{
This function generates the \pkg{lavaan} syntax for a moderated mediation model
with optional moderators on the \emph{a} (\code{x -> m}), \emph{b} (\code{m -> y}), and \emph{c}
(\code{x -> y}) paths. Optionally, covariates can be added, with effects partialled
out from endogeneous (\code{m}, \code{y}) and exogenous (\code{x}, moderators) variables.
}
\details{
This function is a more flexible alternative to the \emph{PROCESS} macro (Hayes, 2022),
which does not require manual specification of a model number. Also, because
the model is estimated in the structural equation modeling framework, both
normal-theory \emph{p}-values and bootstrap intervals are available for indirect effects,
and additional parameters can be computed using the \verb{:=} operator in \pkg{lavaan}.

If no moderators are specified, the model corresponds to the mediation model
provided in \code{\link[=get_mediation_model]{get_mediation_model()}}.

If only one of the \emph{a} and \emph{b} paths is moderated, then indexes of moderated
mediation are computed (labelled with the \code{i_} prefix in the output). If both
the \emph{a} and \emph{b} paths are moderated, indexes of moderated mediation are not
provided (similar to the \emph{PROCESS} macro; see Hayes, 2022).

Currently, the number of \emph{x}, \emph{m}, \emph{y}, and \emph{a}-\emph{b}-\emph{c} moderator variables is
limited to 9 each. This is not a statistical limitation, but instead due to
how the function is written.
}
\examples{
\dontrun{
library(lavaan)
get_moderated_mediation_model(
     x=c("x1", "x2"),
     m=c("m1", "m2"),
     y=c("y1", "y2", "y3"),
     mod_a=c("Sex","Ethnicity"),
     mod_b=NULL,
     mod_c="Sex",
     values_at=list(Sex=c(0,1), Ethnicity=c(0,1)),
     cov=c("IQ","Age")
   ) |>
   sem(df) |>
   summary()
}

}
\references{
Hayes, A. F. (2022).
\emph{Introduction to mediation, moderation, and conditional process analysis} (3rd ed.).
New York: Guilford.
}
\concept{models}
