% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_qca}
\alias{get_qca}
\title{Run qualitative comparative analysis (QCA)}
\usage{
get_qca(
  df,
  dv,
  negate = FALSE,
  conditions,
  n.cut = 1,
  incl.cut = 0.8,
  complex = TRUE,
  dir.exp = NULL,
  ...
)
}
\arguments{
\item{df}{data.frame}

\item{dv}{(character) outcome variable name}

\item{negate}{(logical) negate outcome? (default \code{FALSE})}

\item{conditions}{(character) vector of variable names to use as conditions}

\item{n.cut}{(integer) minimum number of observations in each configuration
(if less than \code{n.cut}, observations are treated as remainders; default 1)}

\item{incl.cut}{(numeric) minimum coherence threshold (default .8, recommended
for fuzzy sets; another conventional value is .75 for crisp sets)}

\item{complex}{(logical) complex solution? (default \code{TRUE}) if \code{FALSE} and
\code{dir.exp} is \code{NULL} (default), the solution is parsimonious,
otherwise solution is intermediate}

\item{dir.exp}{(character) vector indicating directional expectations for each
condition (required for intermediate solution); negate conditions with \code{~}
(e.g. \code{c("cond1", "~cond2")})}

\item{...}{(optional) additional named arguments passed to \code{\link[QCA:truthTable]{QCA::truthTable()}}}
}
\value{
list with elements:
\describe{
\item{nec}{Necessary conditions (from calling \code{\link[QCA:pof]{QCA::pofind()}})}
\item{tt}{Truth table (from calling \code{\link[QCA:truthTable]{QCA::truthTable()}})}
\item{eq}{Minimized solution(s) (from calling \code{\link[QCA:minimize]{QCA::minimize()}})}
}
}
\description{
This function runs a QCA with necessary conditions, truth table, and
minimized solution(s). This is merely a wrapper using the \pkg{QCA} package.
}
\examples{
\dontrun{
df |> get_qca(
   dv="numeric_transformation",
   negate=FALSE,
   conditions=c("RH", "RF", "RI", "RT"),
   n.cut=1,
   incl.cut=.8,
   complex=TRUE
)
}

}
\concept{models}
