% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publications.R
\name{get_sig_effects_from_pub_table}
\alias{get_sig_effects_from_pub_table}
\title{Extract significant effects from a publication table}
\usage{
get_sig_effects_from_pub_table(table, pattern = "\\\\*|\\\\+", fixed = F)
}
\arguments{
\item{table}{Publication generated for lavaan models or tidy summaries}

\item{pattern}{Regular expression used to find significant effects (default is any * or +)}

\item{fixed}{(logical) Whether pattern is to be searched as is (TRUE)
or as a regular expression (FALSE)}
}
\value{
data.frame with columns Outcome, Predictor, and Sign
(indicating whether the coefficient is positive or negative)
}
\description{
This function extracts the significant effects from a publication table generated by any
of the \code{make_pub_table_from_...()} functions.
}
\examples{
library(lavaan)
ana <- list()
ana$y1 <- sem("y1 ~ x1+x2", df)
ana$y2 <- sem("y2 ~ x1+x2", df)
pub <- ana |> make_pub_table_from_lavaan_models()
pub |> get_sig_effects_from_pub_table()

}
\concept{publications}
