% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{group_some_factor_levels}
\alias{group_some_factor_levels}
\title{Combine some factor levels using a map}
\usage{
group_some_factor_levels(df, map)
}
\arguments{
\item{df}{data.frame}

\item{map}{Named list}
}
\value{
\code{df} with factors in \code{names(map)} recoded with combined levels
}
\description{
This function recodes factors by combining some of the original factor levels together.
}
\details{
The names in \code{map} correspond to the factor names in \code{df} whose levels need to be combined.
Each element in \code{map} is a list with one or more vector(s) of factor levels to combine together,
using the first element in each vector as the value for all levels in that vector.
}
\examples{
\dontrun{
map <- list(Nationality=list(c("US", "Puerto Rico"), c("Canada", "Quebec")))
df <- df |> group_some_factor_levels(map)
}
}
