% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{init_h2o}
\alias{init_h2o}
\title{Initialize a local H2O cluster}
\usage{
init_h2o(nthreads = c("half", "minus1", "all"))
}
\arguments{
\item{nthreads}{(character) how many threads (cores) to dedicate to the H2O cluster (see Details)}
}
\value{
New local cluster initialized (return value is given by \code{\link[h2o:h2o.init]{h2o::h2o.init()}})
}
\description{
This function initializes a local H2O cluster for use with the \pkg{h2o}
machine learning library.
}
\details{
The number of available threads is determined using \code{parallel::detectCores()}. The
\code{nthreads} option needs to be one of:
\itemize{
\item \code{half}: Half of all available cores (the default)
\item \code{minus1}: All available threads minus 1 (to avoid freezing)
\item \code{all}: All available cores (\emph{not recommended}; this can freeze your machine)
}

The H2O cluster requires Java to be installed on your machine.
Normally the graphical interface to interact with the cluster is available by
browser at \code{localhost:54321}. Refer to the
\href{https://docs.h2o.ai/h2o/latest-stable/h2o-docs/faq/r.html}{\pkg{h2o} documentation}
for more information.
}
\examples{
\dontrun{
init_h2o() # half cores
init_h2o("minus1") # all but one cores
}

}
\seealso{
\code{\link[h2o:h2o.init]{h2o::h2o.init()}}, \code{\link[parallel:detectCores]{parallel::detectCores()}}
}
