% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{keep_only_vars_in_both_train_and_test}
\alias{keep_only_vars_in_both_train_and_test}
\title{Remove variables from train and test sets when they are unique}
\usage{
keep_only_vars_in_both_train_and_test(tt, remove_from_train_only = FALSE)
}
\arguments{
\item{tt}{train-test list (see \code{\link[=ttsplit]{ttsplit()}})}

\item{remove_from_train_only}{(logical) whether to leave the test set untouched}
}
\value{
\code{tt} with unique variables removed from the train (and potentially test) set(s)
}
\description{
This function removes variables from train and test sets when they
are unique to either the training or test set. Removing variables from the
training set if they cannot be found in the test set is particularly
useful to avoid training a model on information that will not be
available at testing time.
}
\examples{
tt <- df |>
   ttsplit() |>
   keep_only_vars_in_both_train_and_test()

}
\seealso{
\code{\link[=ttsplit]{ttsplit()}}
}
\concept{machine_learning}
