% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logtext}
\alias{logtext}
\title{Log text output to text file}
\usage{
logtext(
  expr,
  logger = get_logger(add_git_info = FALSE, init_log_file = FALSE),
  silent = FALSE,
  return_logger = FALSE
)
}
\arguments{
\item{expr}{R expression(s) to output to text file}

\item{logger}{logger object from \code{\link[=get_logger]{get_logger()}}
(default will create a new logger)}

\item{silent}{(logical) whether to print file path to
console after logging (default \code{TRUE})}

\item{return_logger}{(logical) whether to return the
logger object (useful if no initial logger object was passed)}
}
\value{
Sinks output to text file, and returns logger object if requested
}
\description{
This function outputs (sinks) the console to
a text file on disk.
}
\examples{
\dontrun{
logtext(print("Hello world!"))

logger <- get_logger()
logger |> logtext(expr={
  print("model without covariates:")
  lm(y1 ~ x1, df) |> summary() |> print()
  print("model with covariates")
  lm(y1 ~ x1 + x2 + x3, df) |> summary() |> print()
})
}
}
