% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{make_numeric_if_possible}
\alias{make_numeric_if_possible}
\title{Make numeric if possible}
\usage{
make_numeric_if_possible(df, vars = NULL)
}
\arguments{
\item{df}{data.frame}

\item{vars}{(character) variable names to make numeric (if \code{NULL}, all variables)}
}
\value{
\code{df} with \code{vars} cast into numeric when possible
}
\description{
This function attempts to make all or certain variables numeric.
If a warning or error is thrown, the variable is kept as is.
}
\examples{
df <- df |> make_numeric_if_possible()

}
\concept{data_prep}
