% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publications.R
\name{make_pub_table_from_broom_tidy}
\alias{make_pub_table_from_broom_tidy}
\title{Generate publication table from broom::tidy summaries}
\usage{
make_pub_table_from_broom_tidy(ana)
}
\arguments{
\item{ana}{List of tidy summaries}
}
\value{
data.frame with predictors and coefficients (and significance level) in each model
}
\description{
This function generates a publication table from a list of tidy summaries
to be exported or copy-pasted (e.g. into Excel). The table shows all
coefficients in the 'term' column with significance level:
\itemize{
\item{\code{***}}{p < .001}
\item{\code{**}}{p < .01}
\item{\code{*}}{p < .05}
\item{\code{+}}{p < .10}
}

Each column is a different outcome variable/tidy summary, each row is
a different predictor.
}
\examples{
library(broom)
ana <- list()
ana$y1 <- lm(y1 ~ x1+x2, df) |> tidy()
ana$y2 <- lm(y2 ~ x1+x2, df) |> tidy()
ana |> make_pub_table_from_broom_tidy()
}
