% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{make_pub_table_from_lavaan_models}
\alias{make_pub_table_from_lavaan_models}
\title{Generate publication table from lavaan models}
\usage{
make_pub_table_from_lavaan_models(ana, check_same_format = TRUE)
}
\arguments{
\item{ana}{List of lavaan models}

\item{check_same_format}{(logical) Whether to force all models to have the same predictors}
}
\value{
data.frame with predictors and coefficients (and significance level) in each model
}
\description{
This function generates a publication table from a list of lavaan models
to be exported or copy-pasted (e.g. into Excel). The table shows all
parameters with operator "~" (i.e. regression paths) with significance level:
\itemize{
\item{\code{***}}{p < .001}
\item{\code{**}}{p < .01}
\item{\code{*}}{p < .05}
\item{\code{+}}{p < .10}
}

Each column is a different outcome variable/lavaan model, each row is
a different predictor.
}
\examples{
library(lavaan)
ana <- list()
ana$y1 <- sem("y1 ~ x1+x2", df)
ana$y2 <- sem("y2 ~ x1+x2", df)
ana |> make_pub_table_from_lavaan_models()

}
\concept{models}
