% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{\%P\%}
\alias{\%P\%}
\alias{\%p\%}
\alias{\%c\%}
\title{Pasting operators}
\usage{
lhs \%P\% rhs

lhs \%p\% rhs

lhs \%c\% rhs
}
\arguments{
\item{lhs}{left-hand side}

\item{rhs}{right-hand side}
}
\value{
(character) pasted objects
}
\description{
These operators are used to paste objects together as you would
with the \code{+} operator in Python.
}
\examples{
"this adds a space" \%P\% "between strings"
"this doesn" \%p\% "'" \%p\% "t"
"y1" \%P\% "~" \%P\% (("x" \%p\% 1:3) \%c\% "+") |> formula()
}
\concept{helpers}
