% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{plot_and_print_variable_importances}
\alias{plot_and_print_variable_importances}
\title{Plot and print variable importances from an AutoML model}
\usage{
plot_and_print_variable_importances(automl)
}
\arguments{
\item{automl}{H2O AutoML model (from calling \code{\link[h2o:h2o.automl]{h2o::h2o.automl()}} or \code{\link[=get_automl_model]{get_automl_model()}})}
}
\value{
\code{data.frame} of variable importances
}
\description{
This function plots and prints variable importances from an AutoML
model (if available), and returns variable importances (\code{data.frame}).
}
\examples{
\dontrun{
init_h2o()
tt <- df |> ttsplit()
tt_h2o <- tt |> ship_train_and_test_to_h2o()
config <- get_automl_config()
automl <- config |> get_automl_model(tt_h2o, target="y1")
var_imp <- automl |> plot_and_print_variable_importances()
}

}
\seealso{
\code{\link[h2o:h2o.varimp]{h2o::h2o.varimp()}}, \code{\link[h2o:h2o.varimp_plot]{h2o::h2o.varimp_plot()}}, \code{\link[h2o:h2o.automl]{h2o::h2o.automl()}}, \code{\link[=get_automl_model]{get_automl_model()}}
}
