% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{plot_density_bars_by_groups}
\alias{plot_density_bars_by_groups}
\title{Plot density bars by groups}
\usage{
plot_density_bars_by_groups(df, vars = NULL, group, fix_scales = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{vars}{variables to plot (if \code{NULL}, all variables except \code{group})}

\item{group}{grouping variable}

\item{fix_scales}{(logical) whether to keep same scales for all variables (default \code{FALSE})}
}
\value{
\code{ggplot2} bar plot
}
\description{
This function plots densities using bar graphs, for several variables (panels) separately
by groups (fill color). This is ideal for categorical and ordinal variables or numerical
variables with few categories.
}
\details{
Density is calculated as \code{frequency/total} for each distinct value, separately by groups.
}
\examples{
\dontrun{
df |> plot_density_bars_by_groups(group="Gender")
}
}
