% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{plot_density_by_groups}
\alias{plot_density_by_groups}
\title{Plot density by groups}
\usage{
plot_density_by_groups(
  df,
  vars = NULL,
  group,
  fix_scales = FALSE,
  min_values_to_treat_as_numeric = 12
)
}
\arguments{
\item{df}{data.frame}

\item{vars}{variables to plot (if \code{NULL}, all variables except \code{group})}

\item{group}{grouping variable}

\item{fix_scales}{(logical) whether to keep same scales for all variables (default \code{FALSE})}

\item{min_values_to_treat_as_numeric}{(numeric) number of disctinct values for a
numerical variable to be treated as continuous}
}
\value{
\code{ggplot2} plot
}
\description{
This function plots densities using bar graphs for categorical variables (and actual
densities for numerical variables), for several variables (panels) separately
by groups (fill color). This is ideal for a combination of categorical,
ordinal, and numerical variables, but can also be used with a single type.
}
\details{
Non-numeric variables are always treated as such and get a bar graph. Numeric variables
can get either a bar graph (when the number of unique values excluding NA
is < \code{min_values_to_treat_as_numeric}), or a density plot.

Density for categorical variables is calculated as \code{frequency/total} for each
distinct value, separately by groups.
}
\examples{
\dontrun{
df |> plot_density_by_groups(group="Gender")
}
}
