% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{recode_values_to_NA}
\alias{recode_values_to_NA}
\title{Recode values to NA}
\usage{
recode_values_to_NA(df, vars, values)
}
\arguments{
\item{df}{data.frame}

\item{vars}{(character) vector of variable names}

\item{values}{(vector) values to replace with \code{NA}}
}
\value{
\code{df} with \code{values} in \code{vars} replaced with \code{NA}
}
\description{
This function recodes given values to \code{NA} (missing)
in given variables of a data.frame.
}
\examples{
df <- df |> recode_values_to_NA(vars="x"\%p\%1:3, values=0)

}
\seealso{
\code{\link[=recode_using_excel_map]{recode_using_excel_map()}}
}
\concept{data_prep}
