% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{remove_blank_factor_levels}
\alias{remove_blank_factor_levels}
\title{Remove blank factor levels}
\usage{
remove_blank_factor_levels(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
\code{df} with factors with blank levels replaced with \code{NA} (missing)
}
\description{
This function removes blank factor levels across a data.frame by
replacing blank levels by \code{NA} (missing).
}
\examples{
df <- df |> remove_blank_factor_levels()

}
\concept{data_prep}
