% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{remove_empty_rows}
\alias{remove_empty_rows}
\title{Remove empty rows}
\usage{
remove_empty_rows(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
\code{df} with empty rows removed
}
\description{
This function removes empty rows (all \code{NA}s) from a data.frame
}
\examples{
df <- df |> remove_empty_rows()

}
\concept{data_prep}
