% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{remove_non_ascii_from_df}
\alias{remove_non_ascii_from_df}
\title{Remove non-ASCII characters from data.frame}
\usage{
remove_non_ascii_from_df(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
\code{df} with non-ASCII characters removed
}
\description{
This function removes all non-ASCII characters from a data.frame.
This is particularly useful in non-English locales when special characters
pose problems (e.g. when using the machine learning library \pkg{h2o}).
}
\examples{
df <- df |> remove_non_ascii_from_df()

}
\seealso{
\code{\link[base:iconv]{base::iconv()}}
}
\concept{data_prep}
