% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Flag and remove outliers}
\usage{
remove_outliers(df, vars, rm = TRUE)
}
\arguments{
\item{df}{data.frame}

\item{vars}{(character) variable names used in determining outliers}

\item{rm}{(logical) remove outliers from \code{df} (default TRUE)}
}
\value{
\code{df} with outliers possibly removed
}
\description{
This function flags and removes outliers.
}
\details{
Currently the method implemented uses the Mahalanobis distance to detect outliers,
such that only numeric variables are supported. The critical chi-square value
lies at the 99.9\% mark and is based on the number of variables used in calculating
the distances.

This is the method implemented in \code{\link[psych:outlier]{psych::outlier()}}.
}
\examples{
df <- df |> remove_outliers(vars=c("x1","x2","x3"))

}
\seealso{
\code{\link[psych:outlier]{psych::outlier()}}
}
\concept{data_prep}
