% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{remove_vars_with_too_many_missing}
\alias{remove_vars_with_too_many_missing}
\title{Remove variables with too many missing}
\usage{
remove_vars_with_too_many_missing(df, pmissing = 0.9)
}
\arguments{
\item{df}{data.frame}

\item{pmissing}{(proportion 0-1) maximum proportion of
missingness tolerated in each variable}
}
\value{
\code{df} with variables with too many missing values removed
}
\description{
This function removes variables from a data.frames when their proportion
of missing values is too high.
}
\examples{
df <- df |> remove_vars_with_too_many_missing(pmissing=.80)
}
