% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{replace_in_nested_list}
\alias{replace_in_nested_list}
\title{Replace values in (possibly nested) list of vectors}
\usage{
replace_in_nested_list(list, find, replace, n_max_per_vector = Inf)
}
\arguments{
\item{list}{List of vectors}

\item{find}{Value to replace}

\item{replace}{Replacement value}

\item{n_max_per_vector}{Maximum number of replacements per vector}
}
\value{
List with values replaced
}
\description{
Replace values in (possibly nested) list of vectors
}
\examples{
values <- list(
  c(1,2,3,4,5),
  c(4:10),
  list(
    1:3,
    4:7
  )
)
values |> replace_in_nested_list(find=5, replace=0)

}
\concept{helpers}
