% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{replace_in_vector}
\alias{replace_in_vector}
\title{Replace all occurrences of a value in a vector and replace}
\usage{
replace_in_vector(vec, find, replace, n_max = Inf)
}
\arguments{
\item{vec}{Vector}

\item{find}{Value to replace}

\item{replace}{Replacement value}

\item{n_max}{Maximum number of replacements (starting at index 1)}
}
\value{
Vector with values replaced
}
\description{
Replace all occurrences of a value in a vector and replace
}
\examples{
c(1,1,3,6,6,7,9) |> replace_in_vector(find=6, replace=5)

}
\concept{helpers}
