% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{replace_in_vector_at_position}
\alias{replace_in_vector_at_position}
\title{Replace vector element at given position with value}
\usage{
replace_in_vector_at_position(vec, position, replace)
}
\arguments{
\item{vec}{Vector}

\item{position}{Position at which to replace value}

\item{replace}{Replacement value}
}
\value{
Vector with value replaced
}
\description{
This function replaces the element at a given position of a vector by the replacement value.
}
\examples{
c(1,2,3,4,5) |> replace_in_vector_at_position(1, 0)

}
\concept{helpers}
