% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{rescale_min_max}
\alias{rescale_min_max}
\title{Rescale using min-max normalization}
\usage{
rescale_min_max(df, vars, min = 0, max = 1)
}
\arguments{
\item{df}{data.frame}

\item{vars}{(character) variables to rescale}

\item{min}{(numeric) new minimum value (default 0)}

\item{max}{(numeric) new maximum value (default 1)}
}
\value{
\code{df} with \code{vars} rescaled
}
\description{
This function rescales variable(s) in a data.frame using
min-max normalization with arbitrary min and max values.
}
\details{
Values are rescaled by subtracting the minimum from each value
and multiplying by the new range, and dividing by the current
range, and adding the new minimum (which acts as an offset).
Missing values remain missing. See
\href{https://en.wikipedia.org/wiki/Feature_scaling#Rescaling_(min-max_normalization)}{Feature scaling}
(Wikipedia)
}
\examples{
df <- df |> rescale_min_max(c("x1", "x2", "x3"))

}
\concept{data_prep}
