% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{reverse_code}
\alias{reverse_code}
\title{Reverse-code numeric variables}
\usage{
reverse_code(df, vars)
}
\arguments{
\item{df}{data.frame}

\item{vars}{(character) vector of variable names}
}
\value{
\code{df} with \code{vars} reverse-coded
}
\description{
This function reverse-codes numeric variables. For example,
a 1-7 Likert scale becomes a 7-1 Likert scale (1=7, 2=6, etc.).
This is achieved by taking the maximum value of the variable,
adding the minimum (acting as an offset), and subtracting each value.
}
\examples{
df <- df |> reverse_code("x1")

}
\seealso{
\code{\link[=recode_using_excel_map]{recode_using_excel_map()}}
}
\concept{data_prep}
