% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{run_automl_pipeline}
\alias{run_automl_pipeline}
\title{Run a full AutoML pipeline}
\usage{
run_automl_pipeline(
  df,
  config,
  ...,
  prop_train = 0.7,
  scale_numeric_features = FALSE,
  add_confidence_level = FALSE,
  plot_and_print_variable_importances = TRUE,
  shutdown_h2o = FALSE
)
}
\arguments{
\item{df}{data.frame}

\item{config}{Configuration object (from calling \code{\link[=get_automl_config]{get_automl_config()}})}

\item{...}{additional arguments passed to \code{\link[=get_automl_model]{get_automl_model()}}}

\item{prop_train}{proportion of rows to be allocated to training}

\item{scale_numeric_features}{(logical) whether to scale numeric features
(see \code{\link[=scale_numeric_features_in_train_and_test]{scale_numeric_features_in_train_and_test()}})}

\item{add_confidence_level}{(logical) whether to add confidence levels alongside predictions
in the test set (default \code{FALSE})}

\item{plot_and_print_variable_importances}{(logical) whether to plot and print variable importances
if they are available}

\item{shutdown_h2o}{(logical) whether to shut down cluster at the end of the pipeline
(default \code{FALSE})}
}
\value{
list with elements:
\describe{
\item{tt}{train-test split constructed from \code{df}, with predictions and
possibly confidence levels added to the test set}
\item{automl}{H2O AutoML model}
}
}
\description{
This function runs a full AutoML pipeline, starting from a \code{data.frame}
split into training and test sets to adding predictions from an AutoML model
to the test set and possibly reporting variable importances.
}
\details{
The \code{df} data.frame is first prepared by removing non-ASCII characters,
casting all character variables to factors (if any), and removing blank factor
levels from factors. Then it is split into training and test sets, and if requested,
numeric features are scaled based on sample statistics in the training set. Next,
a local H2O cluster is launched, the train and test sets are uploaded to the cluster,
and an AutoML model is requested using parameters specified in the configuration
object. Predictions are added to the test set (with or without confidence levels),
and variable importances may be plotted and reported. Finally, if requested,
the H2O cluster is shut down.
}
\examples{
\dontrun{
config <- get_automl_config(target="y1")
pip <- df |> run_automl_pipeline(config)
}

}
\seealso{
\code{\link[=ttsplit]{ttsplit()}},
\code{\link[=scale_numeric_features_in_train_and_test]{scale_numeric_features_in_train_and_test()}},
\code{\link[=get_automl_model]{get_automl_model()}},
\code{\link[=add_predictions_from_automl]{add_predictions_from_automl()}}
}
