% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{scale_and_combine}
\alias{scale_and_combine}
\title{Combine numeric variables into one}
\usage{
scale_and_combine(df, vars, name, scale = TRUE)
}
\arguments{
\item{df}{data.frame containing \code{vars}}

\item{vars}{Vector of variable names to be combined into one}

\item{name}{Name of the new variable to be added to df}

\item{scale}{(logical) Whether to scale each variable prior to combining}
}
\value{
df with variable \code{name} added
}
\description{
This function combines vars in \code{vars} (from left to right)
possibly scaling each variable individually prior to combining
}
\examples{
df <- df |> scale_and_combine(c("x1", "x2", "x3"), name="x", scale=TRUE)

}
\concept{data_prep}
