% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{scale_numeric_features_in_train_and_test}
\alias{scale_numeric_features_in_train_and_test}
\title{Scale numeric features in train and test sets}
\usage{
scale_numeric_features_in_train_and_test(tt)
}
\arguments{
\item{tt}{train-test list (see \code{\link[=ttsplit]{ttsplit()}})}
}
\value{
\code{tt} with numeric features scaled
}
\description{
This function scales numeric features in the training set,
and uses the scaling information from the training set
(mean and variance) to scale the corresponding variables
in the test set.
}
\details{
Scaling the test set using variable information in the training set
is performed to avoid target leakage, and to avoid evaluating model
performance on a test set that uses information that might not be available
at test time (i.e. mean and variance of variables for other test observations).
}
\examples{
tt <- df |>
   ttsplit() |>
   scale_numeric_features_in_train_and_test()

}
\seealso{
\code{\link[dataPreparation:build_scales]{dataPreparation::build_scales()}}, \code{\link[=ttsplit]{ttsplit()}}
}
