% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{ship_train_and_test_to_h2o}
\alias{ship_train_and_test_to_h2o}
\title{Send train-test data.frames to H2O cluster}
\usage{
ship_train_and_test_to_h2o(tt)
}
\arguments{
\item{tt}{train-test list (see \code{\link[=ttsplit]{ttsplit()}})}
}
\value{
H2O pointers to the H2O frames (needed to interact with the cluster
using the R API through the \pkg{h2o} package)
}
\description{
This function sends the train and test data.frames to the current
H2O cluster.
}
\details{
Upon running this function, the two data.frames should be
available to the local H2O cluster, and should be accessible
via the graphical interface for the cluster (normally
available through a browser at \code{localhost:54321}) or using
the R API as documented in the \pkg{h2o} package.
}
\examples{
\dontrun{
init_h2o()
tt <- df |> ttsplit()
tt_h2o <- tt |> ship_train_and_test_to_h2o()
}

}
\seealso{
\code{\link[=run_automl_pipeline]{run_automl_pipeline()}}, \code{\link[h2o:as.h2o]{h2o::as.h2o()}}
}
