% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{standardize}
\alias{standardize}
\title{Standardize all numeric variables in a data.frame}
\usage{
standardize(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
\code{df} with all numeric variables (except binary) scaled
}
\description{
This function standardizes (scales) all numeric variables (except binary
variables) in a data.frame. This is done "in place" (no variables are added).
}
\examples{
dfz <- df |> standardize()

}
\concept{data_prep}
