% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{substr_right}
\alias{substr_right}
\title{Substring character from the right side}
\usage{
substr_right(x, n)
}
\arguments{
\item{x}{character}

\item{n}{number of digits to extract starting from the right}
}
\value{
last \code{n}th character of \code{x}
}
\description{
This function returns the nth last characters of a string
}
\examples{
vars <- c("Gender.x", "Gender.y")
vars |> substr_right(1)
}
