% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{tbl_fn_fac}
\alias{tbl_fn_fac}
\title{Internal, generic function to get summary table for factor variables}
\usage{
tbl_fn_fac(df, vars)
}
\arguments{
\item{df}{data.frame}

\item{vars}{Vector of variable names}
}
\value{
Table of descriptive statistics
}
\description{
This function is a generic \code{tbl_fn} function to use with factor variables when calling
\code{\link[=get_desc_table]{get_desc_table()}}.
}
\details{
This should be passed as \code{tbl_fn} when calling \code{\link[=get_desc_table]{get_desc_table()}}. If you want to create a custom
\code{tbl_fn}, you should model it after this one.
}
\examples{
\dontrun{
df |> tbl_fn_fac("Nationality")
}

}
\seealso{
\code{\link[=tbl_fn_num]{tbl_fn_num()}}, \code{\link[=get_desc_table]{get_desc_table()}}
}
\concept{descriptives}
