% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{tbl_fn_num}
\alias{tbl_fn_num}
\title{Internal, generic function to get summary table for numeric variables}
\usage{
tbl_fn_num(df, vars)
}
\arguments{
\item{df}{data.frame}

\item{vars}{Vector of variable names}
}
\value{
Table of descriptive statistics
}
\description{
This function is a generic \code{tbl_fn} function to use with numeric variables when calling
\code{\link[=get_desc_table]{get_desc_table()}}.
}
\details{
This should be passed as \code{tbl_fn} when calling \code{\link[=get_desc_table]{get_desc_table()}}.
If you want to create a custom \code{tbl_fn}, you should model it after this one.
}
\examples{
\dontrun{
df |> tbl_fn_num("age")
}

}
\seealso{
\code{\link[=tbl_fn_fac]{tbl_fn_fac()}}, \code{\link[=get_desc_table]{get_desc_table()}}
}
