% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{to_long}
\alias{to_long}
\title{Pivot data.frame from wide to long}
\usage{
to_long(df, repeated_vars)
}
\arguments{
\item{df}{data.frame}

\item{repeated_vars}{Named list of variables to be pivoted to long format}
}
\value{
\code{df} in long format
}
\description{
This function pivots a data.frame from wide to long.
}
\details{
This is mostly a helper function. For more horsepower, see \code{tidyr::pivot_longer()}.
}
\examples{
v_repeated <- list(
   x=c("x1","x2","x3"),
   y=c("y1", "y2", "y3")
   )
df_long <- df |> to_long(v_repeated)

}
\seealso{
\code{\link[tidyr:pivot_longer]{tidyr::pivot_longer()}}
}
\concept{data_prep}
