% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{transpose_df}
\alias{transpose_df}
\title{Transpose data.frame}
\usage{
transpose_df(df, names_in_first_col = TRUE)
}
\arguments{
\item{df}{data.frame}

\item{names_in_first_col}{(logical) whether to use first column
as new column names (see Details; default \code{TRUE})}
}
\value{
(data.frame) \code{df} transposed
}
\description{
This function transposes a data.frame by switching its rows and columns.
}
\details{
If \code{names_in_first_col} is set to \code{FALSE}, then the first column
in the transposed data.frame will be named \code{Columns}, indicating original column names,
and subsequent columns (original rows) will be named with integers starting at "1".
}
\examples{
df |>
   head() |>
   transpose_df(FALSE)

}
\concept{data_prep}
