% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{ttsplit}
\alias{ttsplit}
\title{Split data.frame into training and test sets}
\usage{
ttsplit(df, prop_train = 0.7)
}
\arguments{
\item{df}{data.frame}

\item{prop_train}{proportion of rows to be allocated to training}
}
\value{
list with elements:
\describe{
\item{train}{\code{df} with \code{prop_train*100}\% rows sampled at random}
\item{test}{\code{df} with remaining rows}
}
}
\description{
This function splits a data.frame into training and test sets.
}
\examples{
tt <- df |> ttsplit()

}
\seealso{
\code{\link[=scale_numeric_features_in_train_and_test]{scale_numeric_features_in_train_and_test()}},
\code{\link[=keep_only_vars_in_both_train_and_test]{keep_only_vars_in_both_train_and_test()}},
\code{\link[=remove_target_from_test_and_add_ref_to_env]{remove_target_from_test_and_add_ref_to_env()}},
\code{\link[=add_target_back_to_test_set_from_ref_table]{add_target_back_to_test_set_from_ref_table()}}
}
\concept{machine_learning}
