% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{view}
\alias{view}
\title{Alias for utils::View()}
\usage{
view(...)
}
\arguments{
\item{...}{Passed to \code{utils::View()}}
}
\value{
Invisible \code{NULL}. See \code{\link[utils:View]{utils::View()}}
}
\description{
This function is mostly useful as a shortcut to avoid default viewer in RStudio.
}
\examples{
\dontrun{
view(df)
}

}
\concept{helpers}
