% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{winsorize}
\alias{winsorize}
\title{Winsorize variables}
\usage{
winsorize(df, var, min = NULL, max = NULL)
}
\arguments{
\item{df}{data.frame}

\item{var}{(character) variable to winsorize}

\item{min}{\code{var}'s new min value (default \code{NULL}, keep as is)}

\item{max}{\code{var}'s new max value (default \code{NULL}, keep as is)}
}
\value{
\code{df} with \code{var} winsorized
}
\description{
This function replaces all values more extreme than a desired
minimum and/or maximum value in a variable by that value.
}
\examples{
df <- df |> winsorize("x1", max=1)

}
\concept{data_prep}
