% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{add_composites}
\alias{add_composites}
\title{Add composites (item means)}
\usage{
add_composites(
  df,
  map,
  na.rm = TRUE,
  standardize_items = FALSE,
  check_already_exist = FALSE
)
}
\arguments{
\item{df}{data.frame}

\item{map}{named list where names are new variable names, elements are
character vectors with variable names to use}

\item{na.rm}{(logical) remove missing (NA) before computing mean?
(default \code{TRUE})}

\item{standardize_items}{(logical) standardize items before computing mean?
(default \code{FALSE})}

\item{check_already_exist}{(logical) stop if variable already exists?
(default \code{FALSE})}
}
\value{
\code{df} with added composite variables
}
\description{
This function adds composites (item means). The Cronbach's alpha is also
printed to the console.
}
\examples{
comp <- list(X=c("x1","x2","x3"))
df <- df |> add_composites(comp)

}
\concept{data_prep}
