% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{copy}
\alias{copy}
\title{Wrapper for writing CSV2 to clipboard}
\usage{
copy(tbl)
}
\arguments{
\item{tbl}{Object to be written}
}
\description{
This is a wrapper for \code{write.csv2()} with \code{file="clipboard-9999"}.
Ideal for copy-pasting a data.frame into Excel in non-English locale.
}
\examples{
df |> copy()

}
\concept{helpers}
