% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\docType{data}
\name{df}
\alias{df}
\title{Simulated cross-lagged dataset}
\format{
A data frame with 500 rows and 6 variables:
\describe{
\item{x1-x3}{\code{x} variable assessed at three occasions}
\item{y1-y3}{\code{y} variable assessed at three occasions}
}
}
\source{
Simulated data using the \code{simsem} package
}
\usage{
df
}
\description{
A simulated cross-lagged dataset with 2 measures (x and y) assessed
three times each (x1-x3, y1-y3)
}
\keyword{datasets}
