% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\docType{data}
\name{df}
\alias{df}
\title{Simulated cross-lagged dataset}
\format{
A data frame with 500 rows and 6 variables:
\describe{
\item{x1}{Numeric variable}
\item{x2}{Numeric variable}
\item{x3}{Numeric variable}
\item{y1}{Numeric variable}
\item{y2}{Numeric variable}
\item{y3}{Numeric variable}
}
}
\source{
Simulated data using the \code{simsem} package
}
\usage{
df
}
\description{
A simulated cross-lagged dataset with 2 measures (x and y) assessed
three times each (x1-x3, y1-y3). The cross-lagged model was generated by
\code{get_crosslagged_model()} and was fed to \code{\link[simsem:generate]{simsem::generate()}} with population
parameter values:
\itemize{
\item \eqn{x_t \sim .7 \times x_{t-1}}
\item \eqn{y_t \sim .7 \times x_{t-1}}
\item \eqn{x_t \sim .4 \times y_{t-1}}
\item \eqn{y_t \sim .4 \times y_{t-1}}
\item \eqn{x_t \sim\sim .35 \times x_t}
\item \eqn{y_t \sim\sim .35 \times y_t}
\item \eqn{y_t \sim\sim .25 \times x_t}
}
}
\concept{datasets}
\keyword{datasets}
