% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gentleman.R
\docType{package}
\name{gentleman}
\alias{gentleman}
\alias{gentleman-package}
\title{gentleman: Helpers for Data Preparation, Descriptives, Models, & Publication}
\description{
This package provides a suite of modules to help with data preparation,
descriptive statistics, modeling, and publication formatting.
The Descriptives module is extensible to accommodate custom functions
to generate tables and p-values (or other useful statistic to compare groups).
The Models module can be used to generate lavaan model strings for mediation
and cross-lagged models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/patc3-phd/gentleman}
}

}
\author{
Patrick Coulombe \email{patrick.coulombe@d22consulting.com}
}
\concept{package}
