% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{get_automl_model}
\alias{get_automl_model}
\title{Run an H2O AutoML model}
\usage{
get_automl_model(tt_h2o, target, config = get_automl_config(), ...)
}
\arguments{
\item{tt_h2o}{list of H2O pointers to H2O frames (see \code{\link[=ship_train_and_test_to_h2o]{ship_train_and_test_to_h2o()}})}

\item{target}{(character) name of the target variable}

\item{config}{configuration object (see \code{\link[=get_automl_config]{get_automl_config()}})}

\item{...}{additional arguments passed to \code{\link[h2o:h2o.automl]{h2o::h2o.automl()}}}
}
\value{
AutoML object from H2O
}
\description{
This function is a wrapper for \code{\link[h2o:h2o.automl]{h2o::h2o.automl()}},
with relevant options specified through the configuration
object generated by \code{\link[=get_automl_config]{get_automl_config()}}.
}
\details{
Only \code{$train} is currently used in \code{tt_h2o}.

If a configuration object is not provided, the default configuration
is the configuration generated by \code{\link[=get_automl_config]{get_automl_config()}}.
}
\examples{
\dontrun{
init_h2o()
tt <- df |> ttsplit()
tt_h2o <- tt |> ship_train_and_test_to_h2o()
automl <- tt_h2o |> get_automl_model(target="y1")
}

}
\seealso{
\code{\link[=get_automl_config]{get_automl_config()}}, \code{\link[h2o:h2o.automl]{h2o::h2o.automl()}}
}
