% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_crosslagged_model}
\alias{get_crosslagged_model}
\title{Generate lavaan syntax for cross-lagged model}
\usage{
get_crosslagged_model(vars_list, random_intercepts = FALSE)
}
\arguments{
\item{vars_list}{List of vectors of variable names of length t (number of repeated
assessments)}

\item{random_intercepts}{(logical) Whether to add random intercepts (RI-CLPM model;
default \code{FALSE})}
}
\value{
Character value to be used with \pkg{lavaan} as model syntax
}
\description{
This function generates the \pkg{lavaan} syntax for a cross-lagged panel model (CLPM).
With random intercepts, the syntax yields the random-intercepts cross-lagged panel
model (RI-CLPM). The function takes an arbitrary number of different variables,
and an arbitrary number of repeated assessments.
}
\details{
The RI-CLPM model requires a minimum of 3 assessments, and is described in
Hamaker et al (2015).
The corresponding model syntax is based on Mulder & Hamaker (2021).

Direct paths are labelled using capital letters for variables and digits for time.
For example, the direct path between the first variable (\code{A}) at time 1 (\code{1})
in \code{vars_list} and the third variable (\code{C}) at time 2 (\code{2}) is labelled \code{A1C2}.
With those labels, you can calculate custom parameters (like indirect effects)
using the \verb{:=} syntax in \pkg{lavaan} (e.g.: \code{ind1 := A1B2*B2C3}).
}
\examples{
library(lavaan)
vars <- list(
   c("x1", "x2", "x3"),
   c("y1", "y2", "y3")
)
get_crosslagged_model(vars) |>
   sem(df) |>
   summary()

}
\references{
Hamaker, E. L., Kuiper, R. M., & Grasman, R. P. P. P. (2015).
A critique of the cross-lagged panel model. \emph{Psychological Methods},
\emph{20}, 102-116.

Mulder, J. D., & Hamaker, E. L. (2021).
Three extensions of the random intercept cross-lagged panel model.
\emph{Structural Equation Modeling: A Multidisciplinary Journal}.
\url{https://doi.org/10.1080/10705511.2020.1784738}.
}
\concept{models}
