% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_mediation_model}
\alias{get_mediation_model}
\title{Generate lavaan syntax for mediation model}
\usage{
get_mediation_model(x, med, y)
}
\arguments{
\item{x}{Vector of predictor (x) variable names}

\item{med}{Vector of mediator (m) variable names}

\item{y}{Vector of outcome (y) variable names}
}
\value{
Character value to be used with \pkg{lavaan} as model syntax
}
\description{
This function generates the \pkg{lavaan} syntax for a mediation model. The model can
take an arbitrary number of predictor (x), mediator (m), and outcome (y) variables.
}
\examples{
library(lavaan)
get_mediation_model("x1", "x2", "x3") |>
   sem(df) |>
   summary()

}
\concept{models}
