% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{get_sig_differences_between_groups}
\alias{get_sig_differences_between_groups}
\title{Get variables for which a group difference exists}
\usage{
get_sig_differences_between_groups(df, test_vars = NULL, group)
}
\arguments{
\item{df}{data.frame}

\item{test_vars}{(character) vector of variable names to test for differences
(if \code{NULL}, all vars)}

\item{group}{(character) name of grouping variable}
}
\value{
(character) vector of test variables for which there is a significant difference
between groups
}
\description{
This function returns a vector of variable names for which
there is a significant difference between a certain group variable.
}
\details{
Group differences are tested using ANOVA (for numeric test variables) or
Chi-square (for categorical test variables) using \code{\link[=get_desc_table]{get_desc_table()}}.
Significant differences are determined using the threshold \code{p < .05}.
}
\examples{
\dontrun{
df |> get_sig_differences_between_groups(c("Age", "Nationality"), group="Gender")
}
}
