% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{\%P\%}
\alias{\%P\%}
\alias{\%p\%}
\alias{\%N\%}
\alias{\%C\%}
\alias{\%c\%}
\title{Pasting operators}
\usage{
lhs \%P\% rhs

lhs \%p\% rhs

lhs \%N\% rhs

lhs \%C\% rhs

lhs \%c\% rhs
}
\arguments{
\item{lhs}{left-hand side}

\item{rhs}{right-hand side}
}
\value{
(character) pasted objects
}
\description{
These operators are used to paste objects together as you would
with the \code{+} operator in Python.
}
\examples{
"this adds a space" \%P\% "between strings"
"this doesn" \%p\% "'" \%p\% "t"
"This skips" \%N\% "a line" |> cat()
"y1" \%P\% "~" \%P\% (("x" \%p\% 1:3) \%C\% "+") # collapse with space
"y1" \%P\% "~" \%P\% (("x" \%p\% 1:3) \%c\% "+") # collapse without space
}
\concept{helpers}
